Update accounts.tsx
I think this condition is correct :) There is discrepancy if "emailVerified" is NULL in database. Even if on User settings -> account page the email stays as verified, on Admin settings -> Users table the icon was not shown.
If user confirm its e-mail address, the "emailVerified" column in database is updated, so "ok" image is displayed correctly in users table and in Account settings page as well.